refactor cli entrypoint into internal package
